+++ src/app/controller/library/browser_controller/helpers.rs
@@ -106,7 +106,7 @@ impl BrowserController<'_> {
             });
         if is_currently_loaded && was_playing {
             let start_override = if playhead_position.is_finite() {
-                Some(playhead_position.clamp(0.0, 1.0))
+                Some(f64::from(playhead_position.clamp(0.0, 1.0)))
             } else {
                 None
             };
@@ -253,7 +253,7 @@ impl BrowserController<'_> {
 
         if is_currently_loaded && was_playing {
             let start_override = if playhead_position.is_finite() {
-                Some(playhead_position.clamp(0.0, 1.0))
+                Some(f64::from(playhead_position.clamp(0.0, 1.0)))
             } else {
                 None
             };
